+++ backend/src/controllers/profileController.js
@@ -5,27 +5,28 @@ setuDb();
 
 exports.create = async (req, res) => {
   try {
+    const refreshToken = req.cookies.refreshToken;
+    if (!refreshToken)
+      return res
+        .status(404)
+        .json({ status: false, message: "You're not logged in!" });
+
     let { mobile, user_id } = req.body;
     const insertData = await Profile.query().insert({
       mobile: mobile,
       user_id: user_id,
     });
 
-    return res.status(201).json({
-      message: "Data berhasil disimpan",
-      data: insertData,
-    });
+    return res.json(insertData);
   } catch (err) {
-    res.json(err);
+    return res.json(err.data);
   }
 };
 
 exports.index = async (req, res) => {
   try {
     const dataProfile = await Profile.query();
-    return res.status(200).json({
-      data: dataProfile,
-    });
+    return res.json({ profiles : dataProfile });
   } catch (err) {
     res.json(err);
   }
@@ -35,13 +36,13 @@ exports.show = async (req, res) => {
   try {
     let id = req.params.id;
     const profile = await Profile.query().findById(id);
+    
     if (!profile) {
-      res.status(404).send({ status: false, message: "Data tidak tersedia!" });
+      res.status(404).send({ message: "Id not Found!" });
       return;
     }
 
     return res.status(200).json({
-      message: "success",
       data: profile,
     });
   } catch (err) {
@@ -50,6 +51,15 @@ exports.show = async (req, res) => {
 };
 
 exports.update = async (req, res) => {
+  const refreshToken = req.cookies.refreshToken;
+  if (!refreshToken)
+    return res
+      .status(404)
+      .json({ status: false, message: "You're not logged in!" });
+
+  const profile = await Profile.query().findById(req.params.id);
+  if (!profile) return res.status(404).json({ message: "Id not found!" });
+
   try {
     let { mobile, user_id } = req.body;
     if (!mobile) return res.json({ message: "Mobile name required!" });
@@ -60,13 +70,9 @@ exports.update = async (req, res) => {
       user_id: user_id,
     });
 
-    if (!profile) {
-      res.status(404).json({ status: false, message: "Data tidak tersedia!" });
-      return;
-    }
-    return res.status(200).json({
-      message: "Data diperbarui!",
-      data: profile,
+    return res.json({
+      message: "Data has change!",
+      data: { profile },
     });
   } catch (err) {
     res.json(err);
@@ -74,14 +80,21 @@ exports.update = async (req, res) => {
 };
 
 exports.destroy = async (req, res) => {
+  const refreshToken = req.cookies.refreshToken;
+  if (!refreshToken)
+    return res
+      .status(404)
+      .json({ status: false, message: "You're not logged in!" });
+
+  const profile = await Profile.query().findById(req.params.id);
+  if (!profile) return res.status(404).json({ message: "Id not found!" });
+  
   try {
     let id = req.params.id;
     const profile = await Profile.query().deleteById(id);
-    if (!profile) return res.status(404).json({ message: "Id not found!" });
     
-    return res.status(200).json({
-      message: "Data berhasil dihapus!",
-      data: profile,
+    return res.json({
+      data: { id },
     });
   } catch (err) {
     res.json(err);
