+++ backend/src/controllers/reviewController.js
@@ -5,18 +5,21 @@ setuDb();
 
 exports.create = async (req, res) => {
   try {
+    const refreshToken = req.cookies.refreshToken;
+    if (!refreshToken)
+      return res
+        .status(404)
+        .json({ status: false, message: "You're not logged in!" });
+
     let { comment, user_id } = req.body;
     const insertData = await Review.query().insert({
       comment: comment,
       user_id: user_id,
     });
 
-    return res.status(201).json({
-      message: "Data berhasil disimpan",
-      data: insertData,
-    });
+    return res.json(insertData);
   } catch (err) {
-    res.json(err);
+    return res.json(err.data);
   }
 };
 
@@ -35,13 +38,13 @@ exports.show = async (req, res) => {
   try {
     let id = req.params.id;
     const review = await Review.query().findById(id);
+    
     if (!review) {
-      res.status(404).send({ status: false, message: "Data tidak tersedia!" });
+      res.status(404).send({ message: "Id not Found!" });
       return;
     }
 
-    return res.status(200).json({
-      message: "success",
+    return res.json({
       data: review,
     });
   } catch (err) {
@@ -50,6 +53,15 @@ exports.show = async (req, res) => {
 };
 
 exports.update = async (req, res) => {
+  const refreshToken = req.cookies.refreshToken;
+  if (!refreshToken)
+    return res
+      .status(404)
+      .json({ status: false, message: "You're not logged in!" });
+
+  const review = await Review.query().findById(req.params.id);
+  if (!review) return res.status(404).json({ message: "Id not found!" });
+
   try {
     let { comment, user_id } = req.body;
     if (!comment) return res.json({ message: "Commant name required!" });
@@ -60,13 +72,9 @@ exports.update = async (req, res) => {
       user_id: user_id,
     });
 
-    if (!review) {
-      res.status(404).json({ status: false, message: "Data tidak tersedia!" });
-      return;
-    }
-    return res.status(200).json({
-      message: "Data diperbarui!",
-      data: review,
+    return res.json({
+      message: "Data has change!",
+      data: { review },
     });
   } catch (err) {
     res.json(err);
@@ -74,14 +82,22 @@ exports.update = async (req, res) => {
 };
 
 exports.destroy = async (req, res) => {
+  const refreshToken = req.cookies.refreshToken;
+  if (!refreshToken)
+    return res
+      .status(404)
+      .json({ status: false, message: "You're not logged in!" });
+
+  const review = await Review.query().findById(req.params.id);
+  if (!review) return res.status(404).json({ message: "Id not found!" });
+
   try {
     let id = req.params.id;
     const review = await Review.query().deleteById(id);
-    if (!review) return res.status(404).json({ message: "Id not found!" });
     
-    return res.status(200).json({
-      message: "Data berhasil dihapus!",
-      data: review,
+    return res.json({
+      message: "Data deleted!",
+      deleted: { id },
     });
   } catch (err) {
     res.json(err);
